% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\docType{data}
\name{grouping}
\alias{grouping}
\title{Grouping of Subjects for the Implicit Association Test}
\format{A data frame with 83 rows and 2 columns.}
\usage{
data(grouping)
}
\description{
Each subject is classified as female (=0) or male (=1).
The variables are as follows:

\itemize{
  \item subId: Subject IDs
  \item gender: Grouping. Female=0 and Male=1
}
}
\examples{
data(grouping)
}
\references{
Lemm, K.M. (2006).
Positive associations among interpersonal contact, motivation, and implicit
and explicit attitudes toward gay men. Journal of Homosexuality 51:2, 79-99.
}
\keyword{datasets}
